+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to the SOTorrent Pipeline project will be documented in this
 ## [Upcoming]
 
 * Compress XML files before upload (use stream compressors such as gzip or bzip2)
-
+* use Python instead of Bash script for bzip2 compression of JSONL files
+* validate checksum, which is returned after upload to Zenodo bucket, using Python instead of Bash script
 ---
 
 The format of this file is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).